Adding "index.html" and images
This is the start of the portfolio, adding the most importante file to this website (because it's the main structure to this project): the index. So, there isn't so much to explain here besides that, like I said, the most importante file to make the website work. To complement, there are some images (for favicon and my own photo) that are being uploaded together with index.